+++ CHANGELOG.md
@@ -22,8 +22,10 @@ in 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
 ## Lab
 
+### 2025-10-21
+  * Added absolute and relative point read from VCB - [discussion](https://forums.sketchup.com/t/explore-the-opencutlist-lab/341006/13)
 ### 2025-10-19
-  * Improved Smart Tools length capture by adding the possibility to use arithmetic operations
+  * Improved Smart Tools length capture by adding the possibility to use arithmetic operations - [discussion](https://forums.sketchup.com/t/explore-the-opencutlist-lab/341006/6)
 ### 2025-10-17
   * Added -@ VCB shortcut
   * Updated codemirror style